requirements.txt
Django==1.8.2
PyYAML==3.11
dj-database-url==0.3.0
django-jsonfield2==15.06.08
django-reversion==1.8.7
django-taggit==0.14.0
waitress==0.8.9
whitenoise==1.0.6
